+++ src/Boot/tests/KernelTest.php
@@ -6,6 +6,8 @@ namespace Spiral\Tests\Boot;
 
 use PHPUnit\Framework\TestCase;
 use Psr\EventDispatcher\EventDispatcherInterface;
+use Spiral\Boot\BootloadManager\Initializer;
+use Spiral\Boot\BootloadManager\InitializerInterface;
 use Spiral\Boot\DispatcherInterface;
 use Spiral\Boot\EnvironmentInterface;
 use Spiral\Boot\Event\Bootstrapped;
@@ -14,6 +16,7 @@ use Spiral\Boot\Event\DispatcherNotFound;
 use Spiral\Boot\Event\Serving;
 use Spiral\Boot\Exception\BootException;
 use Spiral\Core\Container;
+use Spiral\Tests\Boot\Fixtures\CustomInitializer;
 use Spiral\Tests\Boot\Fixtures\TestCore;
 use Throwable;
 
@@ -194,5 +197,26 @@ class KernelTest extends TestCase
         $this->expectException(BootException::class);
 
         $kernel->run()->serve();
+    }
+
+    public function testDefaultInitializerShouldBeBound(): void
+    {
+        $container = new Container();
+
+        TestCore::create(directories: ['root' => __DIR__], container: $container);
+
+        $this->assertTrue($container->has(InitializerInterface::class));
+        $this->assertInstanceOf(Initializer::class, $container->get(InitializerInterface::class));
+    }
+
+    public function testCustomInitializerShouldBeBound(): void
+    {
+        $container = new Container();
+        $container->bind(InitializerInterface::class, CustomInitializer::class);
+
+        TestCore::create(directories: ['root' => __DIR__], container: $container);
+
+        $this->assertTrue($container->has(InitializerInterface::class));
+        $this->assertInstanceOf(CustomInitializer::class, $container->get(InitializerInterface::class));
     }
 }